+++ src/components/property-page-widgets/PropertyPageSearchBar/index.js
@@ -1,0 +1,1 @@
+export { Component as PropertyPageSearchBar } from './component';